pyproject.toml
[project]
name = "terminal-portfolio-optimizer"
version = "1.0.0"
description = "Terminal-based portfolio optimizer for Vietnamese stocks using Textual TUI"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.47.0",
    "vnstock3>=0.1.5",
    "pyportfolioopt>=1.5.5",
    "plotly>=5.18.0",
    "pywebview>=4.4.1",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
]

[project.scripts]
tpo = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
